requirements.txt
aniso8601==4.1.0
atomicwrites==1.3.0
attrs==18.2.0
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.12.0
chardet==3.0.4
Click==7.0
codecov==2.0.15
coverage==4.5.2
coveralls==1.5.1
decorator==4.3.2
docopt==0.6.2
Flask==1.0.2
Flask-RESTful==0.3.7
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==5.0.0
pluggy==0.8.1
psycopg2-binary==2.7.7
py==1.7.0
pycparser==2.19
pytest==4.2.0
pytest-cov==2.6.1
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
validators==0.12.4
Werkzeug==0.14.1
